fix platform not found error
